Add health bar for character
Realized my mistake and changed byte to float

+++ Assets/_Develop/Scripts/Common/Interfaces/IDamageable.cs
@@ -2,6 +2,6 @@ namespace _Develop.Scripts.Enemy
 {
     public interface IDamageable
     {
-        void TakeDamage(byte damage);
+        void TakeDamage(float damage);
     }
 }
